ContractAddressHex to contract address (#118)
* ContractAddressHex -> ContractAddress

# Conflicts:
#	crates/starknet-server/src/api/json_rpc/endpoints.rs

* Compilation fixes

* Moved to types

* Formating applied

* Clippy compilation fixes

* Fixed pr comment

* Apply formatting (import reorder); trigger ci

---------
