+++ crates/starknet-server/src/api/http/endpoints/accounts.rs
@@ -1,11 +1,11 @@
 use axum::extract::Query;
 use axum::{Extension, Json};
+use starknet_types::contract_address::ContractAddress;
 use starknet_types::traits::ToDecimalString;
 
 use crate::api::http::error::HttpApiError;
-use crate::api::http::models::{Balance, ContractAddress, SerializableAccount};
+use crate::api::http::models::{Balance, SerializableAccount};
 use crate::api::http::{HttpApiHandler, HttpApiResult};
-use crate::api::models::ContractAddressHex;
 
 pub(crate) async fn get_predeployed_accounts(
     Extension(state): Extension<HttpApiHandler>,
@@ -19,7 +19,7 @@ pub(crate) async fn get_predeployed_accounts(
         .into_iter()
         .map(|acc| SerializableAccount {
             initial_balance: acc.initial_balance.to_decimal_string(),
-            address: ContractAddressHex(acc.account_address),
+            address: acc.account_address,
             public_key: acc.public_key,
             private_key: acc.private_key,
         })
