+++ crates/starknet-server/src/api/json_rpc/endpoints.rs
@@ -2,6 +2,7 @@ use starknet_core::error::Error;
 use starknet_in_rust::core::errors::state_errors::StateError;
 use starknet_in_rust::transaction::error::TransactionError;
 use starknet_in_rust::utils::Address;
+use starknet_types::contract_address::ContractAddress;
 use starknet_types::felt::{ClassHash, Felt, TransactionHash};
 use starknet_types::starknet_api::block::BlockNumber;
 use starknet_types::traits::ToHexString;
@@ -19,7 +20,7 @@ use crate::api::models::transaction::{
     BroadcastedTransactionWithType, EventFilter, EventsChunk, FunctionCall, Transaction,
     TransactionReceipt, TransactionWithType, Transactions,
 };
-use crate::api::models::{BlockId, ContractAddressHex, PatriciaKeyHex};
+use crate::api::models::{BlockId, PatriciaKeyHex};
 
 /// here are the definitions and stub implementations of all JSON-RPC read endpoints
 impl JsonRpcHandler {
@@ -81,10 +82,7 @@ impl JsonRpcHandler {
             deployed_contracts: state_update
                 .deployed_contracts
                 .into_iter()
-                .map(|(address, class_hash)| DeployedContract {
-                    address: ContractAddressHex(address),
-                    class_hash,
-                })
+                .map(|(address, class_hash)| DeployedContract { address, class_hash })
                 .collect(),
             declared_classes: state_update
                 .declared_classes
@@ -98,16 +96,13 @@ impl JsonRpcHandler {
             nonces: state_update
                 .nonces
                 .into_iter()
-                .map(|(address, nonce)| ContractNonce {
-                    contract_address: ContractAddressHex(address),
-                    nonce,
-                })
+                .map(|(address, nonce)| ContractNonce { contract_address: address, nonce })
                 .collect(),
             storage_diffs: state_update
                 .storage_updates
                 .into_iter()
                 .map(|(contract_address, updates)| StorageDiff {
-                    address: ContractAddressHex(contract_address),
+                    address: contract_address,
                     storage_entries: updates
                         .into_iter()
                         .map(|(key, value)| StorageEntry { key: PatriciaKeyHex(key), value })
@@ -128,7 +123,7 @@ impl JsonRpcHandler {
     /// starknet_getStorageAt
     pub(crate) async fn get_storage_at(
         &self,
-        contract_address: ContractAddressHex,
+        contract_address: ContractAddress,
         key: PatriciaKeyHex,
         block_id: BlockId,
     ) -> RpcResult<Felt> {
@@ -137,7 +132,7 @@ impl JsonRpcHandler {
             .starknet
             .read()
             .await
-            .contract_storage_at_block(block_id.into(), contract_address.0, key.0)
+            .contract_storage_at_block(block_id.into(), contract_address, key.0)
             .map_err(|err| match err {
                 Error::NoBlock => ApiError::BlockNotFound,
                 Error::StateError(StateError::NoneStorage((_, _)))
@@ -193,10 +188,10 @@ impl JsonRpcHandler {
     pub(crate) async fn get_class_hash_at(
         &self,
         block_id: BlockId,
-        contract_address: ContractAddressHex,
+        contract_address: ContractAddress,
     ) -> RpcResult<ClassHash> {
         let starknet = self.api.starknet.read().await;
-        match starknet.get_class_hash_at(&block_id.into(), &contract_address.0) {
+        match starknet.get_class_hash_at(&block_id.into(), &contract_address) {
             Ok(class_hash) => Ok(class_hash),
             Err(Error::NoBlock) => Err(ApiError::BlockNotFound),
             Err(Error::ContractNotFound | Error::NoStateAtBlock { block_number: _ }) => {
@@ -210,7 +205,7 @@ impl JsonRpcHandler {
     pub(crate) async fn get_class_at(
         &self,
         _block_id: BlockId,
-        _contract_address: ContractAddressHex,
+        _contract_address: ContractAddress,
     ) -> RpcResult<ContractClass> {
         Err(error::ApiError::ContractNotFound)
     }
@@ -233,7 +228,7 @@ impl JsonRpcHandler {
         let starknet = self.api.starknet.read().await;
         match starknet.call(
             block_id.into(),
-            request.contract_address.0.into(),
+            request.contract_address.into(),
             request.entry_point_selector,
             request.calldata,
         ) {
@@ -299,14 +294,14 @@ impl JsonRpcHandler {
     pub(crate) async fn get_nonce(
         &self,
         block_id: BlockId,
-        contract_address: ContractAddressHex,
+        contract_address: ContractAddress,
     ) -> RpcResult<Felt> {
         let nonce = self
             .api
             .starknet
             .read()
             .await
-            .contract_nonce_at_block(block_id.into(), contract_address.0)
+            .contract_nonce_at_block(block_id.into(), contract_address)
             .map_err(|err| match err {
                 Error::NoBlock => ApiError::BlockNotFound,
                 Error::NoStateAtBlock { block_number: _ } | Error::ContractNotFound => {
