+++ crates/starknet-server/src/api/models/block.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
+use starknet_types::contract_address::ContractAddress;
 use starknet_types::felt::Felt;
 use starknet_types::starknet_api::block::{BlockNumber, BlockStatus, BlockTimestamp};
 
 use super::transaction::Transactions;
-use super::ContractAddressHex;
 
 pub type BlockHashHex = Felt;
 pub type GlobalRootHex = Felt;
@@ -21,7 +21,7 @@ pub struct BlockHeader {
     pub block_hash: BlockHashHex,
     pub parent_hash: BlockHashHex,
     pub block_number: BlockNumber,
-    pub sequencer_address: ContractAddressHex,
+    pub sequencer_address: ContractAddress,
     pub new_root: GlobalRootHex,
     pub timestamp: BlockTimestamp,
 }
@@ -42,7 +42,7 @@ impl From<&starknet_core::StarknetBlock> for BlockHeader {
             block_hash: value.block_hash(),
             parent_hash: value.parent_hash(),
             block_number: value.block_number(),
-            sequencer_address: ContractAddressHex(value.sequencer_address()),
+            sequencer_address: value.sequencer_address(),
             new_root: value.new_root(),
             timestamp: value.timestamp(),
         }
