+++ crates/starknet-server/src/api/models/mod.rs
@@ -6,15 +6,12 @@ pub(crate) mod transaction;
 
 use serde::{Deserialize, Serialize};
 use starknet_rs_core::types::{BlockId as ImportedBlockId, BlockTag as ImportedBlockTag};
-use starknet_types::contract_address::ContractAddress;
 use starknet_types::felt::Felt;
 use starknet_types::patricia_key::PatriciaKey;
-use starknet_types::starknet_api::block::BlockNumber;
-
-use super::serde_helpers::hex_string::{
-    deserialize_to_prefixed_contract_address, deserialize_to_prefixed_patricia_key,
-    serialize_contract_address_to_prefixed_hex, serialize_patricia_key_to_prefixed_hex,
+use starknet_types::serde_helpers::hex_string::{
+    deserialize_to_prefixed_patricia_key, serialize_patricia_key_to_prefixed_hex,
 };
+use starknet_types::starknet_api::block::BlockNumber;
 
 #[derive(Copy, Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
 pub enum Tag {
@@ -55,16 +52,6 @@ impl From<BlockId> for ImportedBlockId {
         }
     }
 }
-
-/// Contract address serialized/deserialized from/to prefixed hex string
-#[derive(Debug, Default, Copy, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct ContractAddressHex(
-    #[serde(
-        serialize_with = "serialize_contract_address_to_prefixed_hex",
-        deserialize_with = "deserialize_to_prefixed_contract_address"
-    )]
-    pub ContractAddress,
-);
 
 /// Patricia key serialized/deserialized from/to prefixed hex string
 #[derive(Debug, Default, Copy, Clone, Eq, PartialEq, Serialize, Deserialize)]
