+++ crates/types/data/sierra_contract_class_with_abi_as_string.json
