+++ crates/types/src/lib.rs
@@ -4,6 +4,7 @@ pub mod contract_storage_key;
 pub mod error;
 pub mod felt;
 pub mod patricia_key;
+pub mod serde_helpers;
 pub mod traits;
 mod utils;
